fix(test): fix broken tests, and correct an ELS log search bug
